changing the implementation of avoiding race condition in unzip of TRT-LLM wheel by using lock file
